LU-9753 ofd: 64-bits diff variable to avoid overflow
In ofd_create_hdl(), the logic will compare the OST stroed LAST_ID
with the MDT given one: if the difference exceeds some threshold,
then it will trust the OST LAST_ID directly and reset the MDT side
value with the OST one. Otherwise, the orphan OST-objects will be
destroyed.

Unfortunately, both the OST stored LAST_ID and MDT given one are
64 bits, but the @diff variable is only 32 bits, and if the OST
side value is too larger than the MDT side, then the @diff will
overflow. That will misguide the OST to destroy useful OST-objects
by wrong. This patch change the @diff as 64 bits variable.

Reviewed-on: https://review.whamcloud.com/27975
Tested-by: Jenkins
Tested-by: Maloo <hpdd-maloo@intel.com>

+++ lustre/include/lu_target.h
@@ -519,7 +519,7 @@ void tgt_grant_commit(struct obd_export *exp, unsigned long grant_used, int rc);
 int tgt_grant_commit_cb_add(struct thandle *th, struct obd_export *exp,
 			    unsigned long grant);
 long tgt_grant_create(const struct lu_env *env, struct obd_export *exp,
-		      int *nr);
+		      s64 *nr);
 int tgt_statfs_internal(const struct lu_env *env, struct lu_target *lut,
 			struct obd_statfs *osfs, __u64 max_age,
 			int *from_cache);
